requirements.txt
anthropic
cryptography
joblib
kalshi-py
kalshi_python_sync
loguru
openai
pandas
pmxt
plotly
python-dotenv
requests
schedule
scikit-learn
sklearn
streamlit
textblob
winsound
xai_sdk
